+++ packages/__cxstarter__/src/components/Layout/tokens/__cxstarter__Layout.ts
@@ -13,13 +13,15 @@
  */
 
 import { cxLayout, asLayoutToken } from '@bodiless/cx-layout';
+import { __cxstarter__Footer } from '../../Footer';
 import { __cxstarter__Header } from '../../Header';
 
 const Default = asLayoutToken({
   ...cxLayout.Default,
   Components: {
     ...cxLayout.Default.Components,
-    SiteHeader: __cxstarter__Header.Default,
+    Header: __cxstarter__Header.Default,
+    Footer: __cxstarter__Footer.Default,
   },
   Compose: {
     ...cxLayout.Default.Compose,