+++ packages/__cxstarter__/src/styleguide/Page/StyleGuideTemplate/index.tsx
@@ -16,6 +16,9 @@ import { Editors, EditorsMonoFont } from './Editors';
 import { Typography } from './Typography';
 import { Layout } from './Layout';
 import { Header } from './Header';
+import { FlowContainer } from './FlowContainer';
+import { Images } from './Images';
+import { Footer } from './Footer';
 import { _default } from './_default';
 
 export const __cxstarter__StyleGuideTemplate = {
@@ -25,4 +28,7 @@ export const __cxstarter__StyleGuideTemplate = {
   Typography,
   Layout,
   Header,
+  Images,
+  FlowContainer,
+  Footer,
 };