+++ packages/__cxstarter__/src/styleguide/Page/index.tsx
@@ -24,6 +24,9 @@ const {
   Typography,
   Layout,
   Header,
+  FlowContainer,
+  Images,
+  Footer,
   _default
 } = __cxstarter__StyleGuideTemplate;
 
@@ -36,6 +39,9 @@ const Default = asFluidToken({
     Typography,
     Layout,
     Header,
+    FlowContainer,
+    Images,
+    Footer,
   },
 });