+++ packages/cx-layout/src/components/Layout/LayoutClean.tsx
@@ -21,18 +21,19 @@ import {
   A,
 } from '@bodiless/fclasses';
 import { asCxTokenSpec } from '@bodiless/cx-elements';
-// import Footer from '../Footer/Footer';
-import { HeaderClean } from '../Header';
 import { HelmetClean } from '../Helmet';
+import { HeaderClean } from '../Header';
+import { FooterClean } from '../Footer';
 import { LayoutComponents, LayoutProps } from './types';
 
 const layoutComponents: LayoutComponents = {
   OuterContainer: Div,
   SkipToMainContent: A,
   Helmet: HelmetClean,
-  SiteHeader: HeaderClean,
-  // SiteFooter: FooterClean,
-  SiteFooter: Fragment,
+  HeaderWrapper: Fragment,
+  Header: HeaderClean,
+  FooterWrapper: Fragment,
+  Footer: FooterClean,
   Container: Div,
   ContainerWrapper: Div,
   PageTopper: Fragment,
@@ -54,8 +55,10 @@ export const LayoutCleanBase: FC<LayoutProps> = (layoutProps: LayoutProps) => {
     OuterContainer,
     PageCloser,
     PageTopper,
-    SiteFooter,
-    SiteHeader,
+    Header,
+    HeaderWrapper,
+    Footer,
+    FooterWrapper,
     SkipToMainContent,
   } = layoutProps.components;
 
@@ -63,7 +66,9 @@ export const LayoutCleanBase: FC<LayoutProps> = (layoutProps: LayoutProps) => {
     <OuterContainer>
       <SkipToMainContent />
       <Helmet />
-      <SiteHeader />
+      <HeaderWrapper>
+        <Header />
+      </HeaderWrapper>
       <ContainerWrapper>
         <Container>
           <PageTopper />
@@ -71,7 +76,9 @@ export const LayoutCleanBase: FC<LayoutProps> = (layoutProps: LayoutProps) => {
           <PageCloser />
         </Container>
       </ContainerWrapper>
-      <SiteFooter />
+      <FooterWrapper>
+        <Footer />
+      </FooterWrapper>
     </OuterContainer>
   );
 };