+++ packages/cx-layout/src/components/Layout/tokens/cxLayout.ts
@@ -11,31 +11,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import {
   addProps,
   as,
 } from '@bodiless/fclasses';
 import { asLayoutToken } from '../LayoutClean';
+import { cxFooter } from '../../Footer';
 import { cxHeader } from '../../Header';
 import { cxHelmet } from '../../Helmet';
 import { MAIN_CONTENT_ID } from './constants';
-import { WithBordersLabels } from './cxLayoutTest';
+import { StyleGuide } from './StyleGuide';
 
 /**
   * Token that defines a basic layout.
   */
 const Base = asLayoutToken({
   Components: {
-    Helmet: as(cxHelmet.Default),
-  },
-  Theme: {
-    // Tailwind's container is specifially not used due to its feature it set's max-width
-    // to min-width of breakpoint.  So instead rely on ContainerWrapper to margin percent
-    // to contain content.
-    ContainerWrapper: 'mx-site-percent md:mx-md-site-percent lg:mx-lg-site-percent mx-auto',
-    Container: 'mx-auto',
-  },
-  Schema: {
+    Helmet: cxHelmet.Default,
   },
   Behavior: {
     Container: addProps({ id: MAIN_CONTENT_ID }),
@@ -47,37 +40,19 @@ const Base = asLayoutToken({
       'sr-only focus:not-sr-only',
     ),
   },
-  Layout: {
-  },
-  Spacing: {
-  }
-});
-
-const Header = asLayoutToken({
-  Components: {
-    SiteHeader: as(cxHeader.Default),
-  },
-});
-
-const Footer = asLayoutToken({
-  Components: {
-    // SiteFooter: as(cxFooter.Default),
-  },
 });
 
 const Default = asLayoutToken({
   ...Base,
   Components: {
     ...Base.Components,
-    ...Header.Components,
-    ...Footer.Components,
+    Header: cxHeader.Default,
+    Footer: cxFooter.Default,
   },
 });
 
 export default {
   Base,
   Default,
-  Header,
-  Footer,
-  WithBordersLabels,
+  StyleGuide,
 };