+++ packages/cx-navigation/src/components/SubMenu/index.ts
@@ -1,0 +1,3 @@
+export { default as cxSubMenu } from './tokens';
+
+export { asSubMenuToken, asSubMenuWrapperToken } from './SubMenuClean';