+++ packages/cx-navigation/src/components/SubMenu/tokens/index.ts
@@ -1,0 +1,3 @@
+import tokens from './cxSubMenu';
+
+export default tokens;