+++ packages/cx-image/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.2](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/compare/v1.0.0-beta.1...v1.0.0-beta.2) (2022-04-05)
+
+
+### Features
+
+* Add static replacement mechanism ([#1440](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/issues/1440)) ([22384f1](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/commits/22384f184e73366bc00ec619e7b08fa3e5dfc29d))
+* **cx-templates, cx-flowcontainer:** Add generic template  ([#1483](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/issues/1483)) ([d1c1e46](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/commits/d1c1e4652e2be95e3dcf2aba50acf2a6249a1aa9)), closes [#1455](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/issues/1455)
+* **webpack:** Token Shadowing ([#1465](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/issues/1465)) ([1b88de6](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/commits/1b88de677896da149d1d02a8a664bac930d05d53)), closes [#1466](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/issues/1466)
+
+
+
+
+
 # [1.0.0-beta.1](https://sourcecode.jnj.com/scm/asx-nbfp/canvasx/compare/v0.3.7...v1.0.0-beta.1) (2022-03-14)